pyproject.toml
[project]
name = "servicenow-mcp-system"
version = "1.0.0"
description = "ServiceNow MCP integration with Magentic-UI using SSE transport"
requires-python = ">=3.11"
dependencies = [
    # Core MCP and UI
    "fastmcp>=2.11.0",
    "magentic-ui>=0.1.2",
    "mcp[cli]>=1.12.3",
    
    # HTTP and async
    "httpx>=0.27.0",
    "uvicorn>=0.32.1",
    "requests>=2.32.0",
    
    # Environment and configuration
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    
    # ServiceNow integration
    "pysnc>=1.1.8",
    
    # Utilities
    "pathlib2>=2.3.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "*.py",
    "*.yaml", 
    "*.json",
    "openapi_specs/",
]
